src/helm_deploy/requirements.txt
python-dotenv
ska-sdp-config>=0.0.5
ska-sdp-logging>=0.0.5
